requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
datasets==2.9.0
dill==0.3.6
filelock==3.9.0
Flask==2.2.2
Flask-Caching==2.0.2
frozenlist==1.3.3
fsspec==2023.1.0
gunicorn==20.1.0
huggingface-hub==0.12.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
multiprocess==0.70.14
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyarrow==10.0.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
responses==0.18.0
six==1.16.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
xxhash==3.2.0
yarl==1.8.2